backend/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-PyMongo==2.3.0
bcrypt==3.2.0
